Update manifest version to 0.0.2
This commit updates the version in the manifest.json file from 0.0.1 to 0.0.2. This change is necessary to reflect the new version of the SMWS Cask ID replacer extension.

The new version includes minor updates and improvements that are not specified in the issue description but are essential for the ongoing development and maintenance of the extension.

Fixes #SMWS-1